+++ SparkLeMakeDB.sh
@@ -66,8 +66,13 @@ done
 
 # Required Args check:
 # --------------------
-if [ -z "${INPUT_PATH}" ] || [ -z "${OUTPUT_PATH}" ] || [ -z ${HOSTNAME_PREFIX} ]; then
+if [ -z "${INPUT_PATH}" ] || [ -z "${OUTPUT_PATH}" ]; then
     usage
+fi
+
+if [ -z ${HOSTNAME_PREFIX} ] && [ -z ${MASTER_ADDRESS} ]; then
+    echo "Error: Please specify either hostname prefix using --hostnameprefix or Spark master address using --masteraddress"
+    exit 1
 fi
 
 if [ -d "${OUTPUT_PATH}" ]; then
@@ -100,8 +105,8 @@ fi
 # Args correctness checks
 # -----------------------
 
-if [ -z ${SPARK_SLURM_PATH} ]; then
-    echo "Error: Please set SPARK_SLURM_PATH env. variable"
+if [ -z ${SPARK_SLURM_PATH} ] && [ -z ${MASTER_ADDRESS}]; then
+    echo "Error: Please set SPARK_SLURM_PATH env. variable or --masteraddress if a Spark cluster is already up"
     exit 1
 fi
 
@@ -179,7 +184,9 @@ fi
 ${SPARK_HOME}/bin/spark-submit --master ${SPARK_MASTER_ADDRESS} --verbose --executor-cores 1 --class SparkLeMakeDB target/scala-2.11/simple-project_2.11-1.0.jar ${PARTITIONS} ${INPUT_PATH} "${SLB_WORKDIR}/formatdbScript" ${OUTPUT_PATH} ${DB_TYPE};
 
 # Cancel SLURM Job
-HOST_NUMBER=$(echo ${SPARK_MASTER_ADDRESS} | grep -o "[0-9]*" | head -n 1)
-echo ${HOST_NUMBER}
-JOB_ID=$(squeue -u ${USER} | grep ${HOST_NUMBER} | awk '{print $1}')
-scancel ${JOB_ID}
+if [ ! -z ${SPARK_SLURM_PATH} ]; then
+    HOST_NUMBER=$(echo ${SPARK_MASTER_ADDRESS} | grep -o "[0-9]*" | head -n 1)
+    echo ${HOST_NUMBER}
+    JOB_ID=$(squeue -u ${USER} | grep ${HOST_NUMBER} | awk '{print $1}')
+    scancel ${JOB_ID}
+fi
